balluff_master_connector/requirements.txt
Flask
gunicorn
requests
paho-mqtt
flask-mqtt